Double to float

+++ src/com/graphs/lib/graph/element/Point.java
@@ -1,27 +1,27 @@
 package com.graphs.lib.graph.element;
 
 public class Point {
-    private double x;
-    private double y;
+    private float x;
+    private float y;
 
-    public Point(double x, double y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
